fix accept image, video, document in form

+++ routes/teacher.js
@@ -73,8 +73,6 @@ router.post("/edit-quiz/:idLesson", (req, res) => {
         })
 })
 
-
-
 //TODO: dashboard teacher
 router.get("/dashboard", async (req, res) => {
     var infoTeacher='';
@@ -116,7 +114,7 @@ router.get("/add-course", (req, res) => {
         res.redirect('/login')
     } else {
         let getInfoTeacher = jwtDecode(req.session.token)
-        if(getInfoTeacher == "Teacher"){
+        if(getInfoTeacher.role == "Teacher"){
             res.render("teacher/instructor-add-course", {
                 teacher: getInfoTeacher,
             });
@@ -132,13 +130,14 @@ router.post("/add-course", upload, async (req, res) => {
         let getInfoTeacher = jwtDecode(req.session.token);
         if (getInfoTeacher.role == "Teacher"){
             let formData = await new FormData();
-            let readStream = fs.createReadStream(`./public/upload/tmp/${req.files.image.originalname}`);
+            let readStream = fs.createReadStream(`./public/upload/tmp/${req.files.image[0].originalname}`);
 
             const formHeaders = formData.getHeaders();
             formData.append("name", req.body.name);
             formData.append("topic", req.body.topic);
             formData.append("description", req.body.description);
             formData.append("price", req.body.price);
+            formData.append("tag", req.body.tag);
             formData.append("image", readStream);
             // console.log(formData)
             let config_axios = {
@@ -148,6 +147,65 @@ router.post("/add-course", upload, async (req, res) => {
                 }
             };
             await axios.post(`${domain}/api/course`, formData, config_axios)
+                .then(function (response) {
+                    fs.unlink(`/public/tmp/${req.files.image[0].originalname}`);
+                    // res.send(response)
+                    res.redirect("/teacher/courses")
+                })
+                .catch(function (error) {
+                    res.send(error)
+                });
+        } else{
+            res.redirect('/')
+        }
+    } else{
+        res.redirect('/login');
+    }
+});
+
+router.get("/edit-course/:idCourse", async (req, res) => {
+    if(req.session.token){
+        let getInfoTeacher = jwtDecode(req.session.token);
+        await axios({
+            method: 'get',
+            url: `${domain}/api/course/${req.params.idCourse}`
+        }).then(result => {
+            if(result.data.teacher == getInfoTeacher._id){
+                res.render('teacher/instructor-edit-course', {course: result.data});
+            }else{
+                 res.redirect('/course');
+            }
+        }).catch(error => {
+            res.send(error.message)
+        })
+    }else{
+        res.redirect('/login');
+    }
+});
+
+router.post("/edit-course/:idCourse", async (req, res) => {
+    if (req.session.token) {
+        let getInfoTeacher = jwtDecode(req.session.token);
+        res.send(getInfoTeacher)
+        if (getInfoTeacher.role == "Teacher"){
+            let formData = await new FormData();
+            let readStream = fs.createReadStream(`./public/upload/tmp/${req.files.image.originalname}`);
+
+            const formHeaders = formData.getHeaders();
+            formData.append("name", req.body.name);
+            formData.append("topic", req.body.topic);
+            formData.append("description", req.body.description);
+            formData.append("price", req.body.price);
+            formData.append("tag", req.body.tag);
+            formData.append("image", readStream);
+            console.log(formData)
+            let config_axios = {
+                headers: {
+                    Authorization: req.session.token,
+                    ...formHeaders
+                }
+            };
+            await axios.put(`${domain}/api/course/${req.params.idCourse}`, formData, config_axios)
                 .then(function (response) {
                     fs.unlink(`/public/tmp/${req.file.originalname}`);
                     res.send(response)
@@ -162,57 +220,48 @@ router.post("/add-course", upload, async (req, res) => {
     } else{
         res.redirect('/login');
     }
-});
-
-router.get("/edit-course/:idCourse", (req, res) => {
-    axios({
-        method: 'get',
-        url: `${domain}/api/course/${req.params.idCourse}`
-    }).then(result => {
-        res.render('teacher/instructor-edit-course', {course: result.data});
-    }).catch(error => {
-        res.send(error.message)
-    })
-    
 })
+
+router.post("/add-lesson/:idCourse", upload, async (req, res) => {
+    // res.json(req.files) 
+    if(req.session.token){
+        for(let i = 0; i < req.body.title.length; i++){
+            let formData = await new FormData();
+            let readStreamVideo = fs.createReadStream(`./public/upload/tmp/${req.files.video[i].originalname}`);
+            let readStreamDoc = fs.createReadStream(`./public/upload/tmp/${req.files.document[i].originalname}`);
+    
+            const formHeaders = formData.getHeaders();
+            formData.append("title", req.body.title[i]);
+            formData.append("description", req.body.description[i]);
+            formData.append("video", readStreamVideo);
+            formData.append("document", readStreamDoc);
+            // console.log(formData)
+            let config_axios = {
+                headers: {
+                    Authorization: req.session.token,
+                    ...formHeaders
+                }
+            };
+            await axios.post(`${domain}/api/lesson/${req.params.idCourse}`, formData, config_axios)
+                .then(function (response) {
+                    fs.unlink(`/public/tmp/${req.files.video[i].originalname}`);
+                    fs.unlink(`/public/tmp/${req.files.document[i].originalname}`);
+                    console.log(`created successfully for ${req.body.title[i]}`);
+                })
+                .catch(function (error) {
+                    res.send(error)
+                });
+        }
+        res.redirect("/teacher/courses");
+    } else {
+        res.redirect('/login');
+    }
+})
+
 
 router.post("/edit-lesson/:idCourse", upload ,async (req, res) => {
     
 })
-
-
-router.post("/add-lesson/:idCourse", upload ,async (req, res) => {
-    // res.json(req.files)
-    for(let i = 0; i < req.body.title.length; i++){
-        let formData = await new FormData();
-        let readStreamVideo = fs.createReadStream(`./public/upload/tmp/${req.files.video[i].originalname}`);
-        let readStreamDoc = fs.createReadStream(`./public/upload/tmp/${req.files.document[i].originalname}`);
-
-        const formHeaders = formData.getHeaders();
-        formData.append("title", req.body.title[i]);
-        formData.append("description", req.body.description[i]);
-        formData.append("video", readStreamVideo);
-        formData.append("document", readStreamDoc);
-        // console.log(formData)
-        let config_axios = {
-            headers: {
-                Authorization: req.session.token,
-                ...formHeaders
-            }
-        };
-        await axios.post(`${domain}/api/lesson/${req.params.idCourse}`, formData, config_axios)
-            .then(function (response) {
-                fs.unlink(`/public/tmp/${req.files.video[i].originalname}`);
-                fs.unlink(`/public/tmp/${req.files.document[i].originalname}`);
-                console.log(`created successfully for ${req.body.title[i]}`);
-            })
-            .catch(function (error) {
-                res.send(error)
-            });
-    }
-    res.redirect("/teacher/courses");
-})
-
 
 router.get("/courses", (req, res) => {
     let getInfoTeacher = jwtDecode(req.session.token)
